Copter: Move send_*(mavlink_channel chan) to GCS_MAVLINK_Copter
These methods were only serving as a way for GCS_MAVLINK_Copter objects
to 'ask' the Copter to fulfill the mavlink messages with data and send it.
But the odd thing is that those methods were triggered by the Copter asking
them to send those messages.

Since GCS_MAVLINK_Copter objects are friends of Copter, they can fulfill
and send the messages by themselves (it becomes an inner call instead
cross objects).

* This avoid part of the excessive ping-pong between Copter and
  GCS_MAVLink_Copter when sending MAVLink messages
  * There is still room for improvements
* Will help simplify the code so we can factory good part of the common
  code between vehicles in a near future