Vehicles: Make a more noble use of AP_Vehicle
There are a very large number of quadruplicated code among ArduPilot
vehicles implementation. Much of this code can be easily shared and some
of the differences can be exposed through a common API taking advantage
of Object orientation.

At the same time AP_Vehicle class was only being used to exposed two
public structures. This patch promotes this class to be the common
ground between ArduPilot vehicles and provides a place to hold all the
common code between them. Upon completion, the two parameters structures
being hold by this class will also move for a better place, but it won't
hurt to keep it there for now.